+++ lib/main.dart
@@ -210,7 +210,7 @@ class PeercoinApp extends StatelessWidget {
               child: SpinningPeercoinIcon(),
             ),
             child: MaterialApp(
-              title: 'Peercoin',
+              title: 'Sumcoin',
               debugShowCheckedModeBanner: false,
               supportedLocales: AppLocalizations.availableLocales.keys
                   .map((lang) => Locale(lang)),