+++ test_driver/key_new.dart
@@ -85,8 +85,8 @@ void main() {
           await driver.runUnsynchronized(
             () async {
               await driver.tap(find.byValueKey('newWalletIconButton'));
-              await driver.tap(find.text('Peercoin Testnet'));
-              await driver.tap(find.text('Peercoin Testnet')); //tap into wallet
+              await driver.tap(find.text('Sumcoin Testnet'));
+              await driver.tap(find.text('Sumcoin Testnet')); //tap into wallet
               expect(await driver.getText(find.text('connected')), 'connected');
             },
           );